+++ packages/terra-dropdown-button/src/_Dropdown.jsx
@@ -29,7 +29,7 @@ const propTypes = {
   /**
    * Whether or not dropdown is opened using keyboard.
    */
-  isKeyboardEvent: PropTypes.bool,
+  openedViaKeyboard: PropTypes.bool,
   /**
    * Callback for reference of the dropdown button
    */
@@ -41,7 +41,7 @@ const propTypes = {
 };
 
 const Dropdown = ({
-  requestClose, isOpen, targetRef, children, width, isKeyboardEvent, buttonRef, refCallback,
+  requestClose, isOpen, targetRef, children, width, openedViaKeyboard, buttonRef, refCallback,
 }) => (
   <Hookshot
     isOpen={isOpen}
@@ -55,7 +55,7 @@ const Dropdown = ({
       onEsc={requestClose}
       onOutsideClick={requestClose}
     >
-      <FocusTrap focusTrapOptions={{ returnFocusOnDeactivate: true, initialFocus: isKeyboardEvent ? '' : buttonRef, clickOutsideDeactivates: true }}>
+      <FocusTrap focusTrapOptions={{ returnFocusOnDeactivate: true, initialFocus: openedViaKeyboard ? '' : buttonRef, clickOutsideDeactivates: true }}>
         <DropdownList
           requestClose={requestClose}
           width={width}