+++ packages/terra-dropdown-button/src/_DropdownList.jsx
@@ -149,6 +149,8 @@ class DropdownList extends React.Component {
     return React.Children.map(this.props.children, (child, index) => React.cloneElement(child, {
       isActive: index === this.state.active,
       requestClose: this.props.requestClose,
+      ...(index === 0) && { 'data-terra-dropdown-first-list-item': true },
+      ...(index === this.props.children.length - 1) && { 'data-terra-dropdown-last-list-item': true },
     }));
   }